"avance"

+++ sprint/dentalClinic/dentalClinic/src/main/java/com/proyectoIntegrador/dentalClinic/dto/salida/paciente/DomicilioSalidaDto.java
@@ -12,7 +12,7 @@ import javax.validation.constraints.Pattern;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class DomicilioSalidaDto {
 
-    private int id;
+    private Long id;
     private String calle;
     private int numero;
     private String localidad;
@@ -21,7 +21,7 @@ public class DomicilioSalidaDto {
     public DomicilioSalidaDto() {
     }
 
-    public DomicilioSalidaDto(int id, String calle, int numero, String localidad, String provincia) {
+    public DomicilioSalidaDto(Long id, String calle, int numero, String localidad, String provincia) {
         this.id = id;
         this.calle = calle;
         this.numero = numero;
